Add some more small helpers

+++ include/meow_hook/util.h
@@ -5,7 +5,8 @@
 
 namespace meow_hook
 {
-template <typename R> static R __thiscall func_call(uint64_t addr)
+
+template <typename R> static R __thiscall func_call(uintptr_t addr)
 {
     return ((R(*)())(addr))();
 }
@@ -20,7 +21,7 @@ template <typename R, typename... Args> struct func_call_member_helper {
 };
 
 template <typename R, class T, typename... Args>
-R func_call_member(uint64_t addr, T _this, Args... args)
+R func_call_member(uintptr_t addr, T _this, Args... args)
 {
     const func_call_member_helper<R, Args...> *helper =
         reinterpret_cast<const func_call_member_helper<R, Args...> *>(_this);
@@ -30,15 +31,16 @@ R func_call_member(uint64_t addr, T _this, Args... args)
 
 template <typename R, class T, typename... Args>
 inline std::enable_if_t<!std::is_pointer_v<T> || !std::is_class_v<std::remove_pointer_t<T>>, R>
-func_call(uint64_t addr, T _this, Args... args)
+func_call(uintptr_t addr, T _this, Args... args)
 {
     return ((R(*)(T, Args...))(addr))(_this, args...);
 }
 
 template <typename R, class T, typename... Args>
 inline std::enable_if_t<std::is_pointer_v<T> && std::is_class_v<std::remove_pointer_t<T>>, R>
-func_call(uint64_t addr, T _this, Args... args)
+func_call(uintptr_t addr, T _this, Args... args)
 {
     return func_call_member<R, T, Args...>(addr, _this, args...);
 }
+
 } // namespace meow_hook
